+++ include/Cola.h
@@ -14,7 +14,7 @@ typedef struct _cola cola;
 cola *cola_vacia ();
 bool es_vacia (cola*);
 void borrar_cola (cola*);
-void encolar (cola *, char, int);
+void encolar (cola *, char, float);
 void desencolar (cola *);
 int primero (cola *);
 void imprimir (cola *);
